fix: nfsv3driver: run rpcbind / rpc.statd
Mounting NFS volumes requires rpc.statd to be running (for lock
management).  As we run each job in a separate container, we cannot just
add jobs for them (as then rpc.statd would not be able to communicate
with rpcbind correctly).  Instead, just spawn them in the nfsv3driver
startup script; they will background themselves.

Also, remove the logging redirects, as logging to stdout/stderr is more
useful in a containerized scenario.