fix: nfsv3driver: rename pre-start script
If the pre-start script is named bin/pre-start, it will (also) be run as
a BOSH pre-start script (in addition to being run as a BPM pre-start
script); quarks-operator will in this case mount
/var/vcap/data/voldrivers as read-only, which breaks the pre-start
script.  Only use it as a BPM pre-start script instead to ensure that
the volume is mounted read-write.